Handle const and pointer arguments and return types (#14)
The CMock header parser does some complicated stuff with const and pointer arguments and return values. Also, the way this is done has changed recently.

These changes improve the way const and pointer arguments are handled so that fake functions should always be generated correctly. The logic for function arguments is also applied to function return types to fix issues with them as well.

This change also makes our own copy of the CMock header parser in the plugin (in a new filename: fff_header_parser) so that changes won't affect this plugin.

Fixes #11.
Fixes #3.